requirements.txt
# Core dependencies
telethon
pydantic
pandas
numpy
python-dotenv

# Logging
structlog

# Database
psycopg2-binary
sqlalchemy

# Utilities
tqdm
aiofiles

# API
fastapi
uvicorn

# Computer Vision
ultralytics
opencv-python

# Orchestration
dagster
dagster-postgres

# Testing
pytest
pytest-asyncio

 ultralytics 
 opencv-python 
 pillow


 fastapi 
 uvicorn 
 sqlalchemy 
 psycopg2-binary 
 python-multipart

 dagster 
 dagster-webserver 
 dagster-postgres 
 pandas 
 psycopg2-binary